requirements.txt
faiss-cpu
pyyaml
requests
tiktoken

